SPK excerpt generator (#13)
SPK can be large files, especially if they cover a large range of time.

Take `de441`, it covers from -13200 to 17191 and its size is 3.1 GB.
What if you need data for year 3000, and only for the Sun? You would need to deal with this massive ephemeris which takes a long time to be parsed and used.

This introduces the ability to create an excerpt from an original SPK by reducing the targets and/or the time range.

```
$ ruby-ephem excerpt --targets 10 2100-01-01 2101-01-01 de430.bsp excerpt.bsp
```

`de430.bsp` is 120 MB while `excerpt.bsp` is 12 KB.

This should be particularly convenient when you are limited in the targets you need and the years you want to support.