IMCCE INPOP support
I can't hide this change makes me super happy.

_INPOP_ is the ephemerides series from the Paris Observatory, more
precisely from the IMCCE laboratory. The Paris Observatory has been a
central place for astronomy for the last few centuries and the data
produces today by the IMCCE is among the most accurate and precise we
can produce today.

I frequently use data from the IMCCE in the development and tests of
Astronoby, main library using Ephem at the moment, so it feels right to
support their ephemerides as well.

I am extremely lucky that they offer a SPICE version of their ephemeris
files for most of INPOP's versions. As you can see, this change doesn't
implement any modification of the SPK parsing logic, it only adds the
support for downloading files from somewhere else than the JPL.

The choice between JPL and IMCCE is just a matter of preference, their
data being extremely close. While you can witness differences of a few
kilometers in positions, the difference becomes barely noticeable when
used in libraries transforming the data into topocentric coordinates
like Astronoby.

The list of supported files is:
* `inpop10b.bsp`
* `inpop10b_large.bsp`
* `inpop10e.bsp`
* `inpop10e_large.bsp`
* `inpop13c.bsp`
* `inpop13c_large.bsp`
* `inpop17a.bsp`
* `inpop17a_large.bsp`
* `inpop19a.bsp`
* `inpop19a_large.bsp`
* `inpop21a.bsp`
* `inpop21a_large`

The difference between ephemerides with or without `_large` is the time
period: from 1900 to 2100 for regular ephemerides, from 1000 to 3000,
Common Era.