Validate against all kernels and date ranges (#36)
Before, only one kernel and one date range was tested when checking the
accuracy of Ephem. This was due to the CI time it used to take.

Now that the performance have been greatly improved, these tasks are way
faster. Therefore, it is more affordable to check the accuracy against
all the available data.